test: Add unit tests for weighted attributed directed graph

+++ test/weighted_hetero_digraph_test.cpp
@@ -28,7 +28,7 @@ TEST(WeightedHeteroDigraphTest, SetNode) {
     graph.setNode(1);
     graph.setNode(2);
 
-    EXPECT_EQ(graph.size(), 3);
+    EXPECT_EQ(graph.size(), static_cast<size_t>(3));
 }
 
 TEST(WeightedHeteroDigraphTest, RemoveNode) {
@@ -38,7 +38,7 @@ TEST(WeightedHeteroDigraphTest, RemoveNode) {
     graph.setNode(2);
 
     graph.removeNode(1);
-    EXPECT_EQ(graph.size(), 2);
+    EXPECT_EQ(graph.size(), static_cast<size_t>(2));
 }
 
 TEST(WeightedHeteroDigraphTest, AddEdge) {
@@ -86,7 +86,7 @@ TEST(WeightedHeteroDigraphTest, GetAdjacents) {
     for (const auto& [id, weight] : adjacents) {
         spdlog::debug("id: {}, weight: {}", id, weight);
     }
-    EXPECT_EQ(adjacents.size(), 2);
+    EXPECT_EQ(adjacents.size(), static_cast<size_t>(2));
     EXPECT_DOUBLE_EQ(adjacents.at(1), 5.0);
     EXPECT_DOUBLE_EQ(adjacents.at(2), 2.5);
 }
@@ -108,7 +108,7 @@ TEST(WeightedHeteroDigraphTest, GetSubgraph) {
     WeightedHeteroDigraph<int> subgraph = graph.getSubgraph(indices);
 
     spdlog::debug("calculate subgraph size");
-    EXPECT_EQ(subgraph.size(), 3);
+    EXPECT_EQ(subgraph.size(), static_cast<size_t>(3));
     EXPECT_DOUBLE_EQ(subgraph.getWeight(0, 1), 5.0);
     EXPECT_DOUBLE_EQ(subgraph.getWeight(1, 3), 3.0);
 
@@ -126,7 +126,7 @@ TEST(WeightedHeteroDigraphTest, Organize) {
     graph.addEdge(2, 4, 3.5);
 
     graph.organize();
-    EXPECT_EQ(graph.size(), 3);
+    EXPECT_EQ(graph.size(), static_cast<size_t>(3));
     EXPECT_DOUBLE_EQ(graph.getWeight(0, 1), 5.0);
     EXPECT_DOUBLE_EQ(graph.getWeight(1, 2), 3.5);
 
@@ -191,7 +191,7 @@ TEST(WeightedHeteroDigraphTest, ReadGraph) {
     WeightedHeteroDigraph<int> graph;
     graph.readGraph("../../dataset/graph.txt", FileExtension::TXT);
 
-    EXPECT_EQ(graph.size(), 6);
+    EXPECT_EQ(graph.size(), static_cast<size_t>(6));
     EXPECT_EQ(graph.getWeight(0, 1), 1.0);
     EXPECT_EQ(graph.getWeight(0, 2), 1.0);
     EXPECT_EQ(graph.getWeight(1, 2), 1.0);
